refactor: [searchResults] 관련 타입 지정
